Fixed lcl reset time unsigned issue.

+++ src/cons/cons.hpp
@@ -85,7 +85,7 @@ struct consensus_context
     //ledger close time of previous hash
 
     uint64_t prev_close_time = 0;
-    uint64_t reset_time = 0;
+    uint16_t reset_time = 0;
 
     bool is_state_syncing = false;
     std::string state_sync_lcl;
